requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-adminlte-3==0.1.6
django-cors-headers==3.10.0
django-filter==21.1
django-import-export==2.6.1
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
# pkg_resources==0.0.0
PyJWT==1.7.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
typing-extensions==3.10.0.2
xlrd==2.0.1
xlwt==1.3.0
